+++ challenge01/src/App.jsx
@@ -1,26 +1,15 @@
 import "./App.css";
 import { useState } from "react";
 import List from "./components/List";
-import { ImageSizeContext } from "./contexts/ImageSizeContext";
+import ImageSizeProvider from "./contexts/ImageContext/ImageSizeProvider";
+import ImageSizeToggle from "./components/ImageSizeToggle";
 function App() {
-  const [isLarge, setIsLarge] = useState(false);
-  const imageSize = isLarge ? 150 : 100;
-
   return (
     <div className="App">
-      <ImageSizeContext.Provider value={imageSize}>
-        <label>
-          <input
-            type="checkbox"
-            checked={isLarge}
-            onChange={(event) => {
-              setIsLarge(event.target.checked);
-            }}
-          />
-          Use large images
-        </label>
-        <List imageSize={imageSize} />
-      </ImageSizeContext.Provider>
+      <ImageSizeProvider>
+        <ImageSizeToggle />
+        <List />
+      </ImageSizeProvider>
     </div>
   );
 }
